+++ src/utils/output_json.h
@@ -38,6 +38,9 @@ rapidjson::Value to_json(const Step& s,
 rapidjson::Value to_json(const CostBreakdown& breakdown,
                          rapidjson::Document::AllocatorType& allocator);
 
+rapidjson::Value to_json(const UnassignedDetails& details,
+                         rapidjson::Document::AllocatorType& allocator);
+
 rapidjson::Value to_json(const Location& loc,
                          rapidjson::Document::AllocatorType& allocator);